requirements.txt
Flask
Flask-Script
Flask-Restplus
selenium
Flask-CORS
Flask-SQLAlchemy
requests

# Also needs: phobos
